Undo/Redo support.  Very risky code since Unity does not provide an undo/Redo callback…
Currently  relies on saving a boolean IsUndoRedoPerformed with a true value before performing a snapshot.  The IStorage.Update that follows an Undo/Redo operation will resynchronize the internal structures once the Undo/Redo is detected.  However, there is no warranty that an other IStorage function gets called before the update creating a mess….

Signed-off-by: Reinual <michel.launier@videotron.ca>